papDjango/requirements.txt
django-mssql-backend
django-filter
django-crispy-forms 
django-tables2
djangorestframework
ODBC Driver 13 for SQL Server